==> analysis/07-cluster-analysis-all-timepoints.Rmd <==
---
title: "Cluster analysis of timeline experiment"
author: "Katharina Hembach"
date: "11/10/2020"
output: 
  html_document:
    toc: true,
    code_folding: show
---
```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE, autodep = TRUE, cache = TRUE, dev = "png",
                      dev.args = list(png = list(type = "cairo")), 
                      message = FALSE, cache.lazy = FALSE)
```


### Load packages

```{r load-libs, message = FALSE, warning = FALSE}
library(ComplexHeatmap)
library(cowplot)
library(ggplot2)
library(dplyr)
library(muscat)
library(purrr)
library(RColorBrewer)
library(viridis)
library(scran)
library(Seurat)
library(SingleCellExperiment)
library(stringr)
library(RCurl)
library(BiocParallel)
```

# Load data & convert to SCE

```{r load-data}
so <- readRDS(file.path("output", "so_06-clustering_all_timepoints.rds"))
sce <- as.SingleCellExperiment(so, assay = "RNA")
colData(sce) <- as.data.frame(colData(sce)) %>% 
    mutate_if(is.character, as.factor) %>% 
    DataFrame(row.names = colnames(sce))
```

# Number of clusters by resolution

```{r}
cluster_cols <- grep("res.[0-9]", colnames(colData(sce)), value = TRUE)
sapply(colData(sce)[cluster_cols], nlevels)
```

# Cluster-sample counts

```{r}
# set cluster IDs to resolution 0.4 clustering
so <- SetIdent(so, value = "RNA_snn_res.0.4")
so@meta.data$cluster_id <- Idents(so)
sce$cluster_id <- Idents(so)
(n_cells <- table(sce$cluster_id, sce$sample_id))
so$group_id <- factor(so$group_id, levels = c("P22", "D52", "D96", "D223"))
```

# Relative cluster-abundances

```{r fig.width = 6, fig.height = 5}
fqs <- prop.table(n_cells, margin = 2)
mat <- round(as.matrix(unclass(fqs))*100, 2)
colfunc <- colorRampPalette(c("ghostwhite", "deepskyblue4"))
Heatmap(mat,
    col = colfunc(10),
    name = "Percentage\nof cells",
    cluster_rows = FALSE,
    cluster_columns = FALSE,
    row_names_side = "left",
    row_title = "cluster ID",
    column_title = "sample ID",
    column_title_side = "bottom",
    rect_gp = gpar(col = "white"),
    cell_fun = function(i, j, x, y, width, height, fill)
        grid.text(mat[j, i], x = x, y = y, 
            gp = gpar(col = "black", fontsize = 8)))
```


# Cell cycle scoring with Seurat

We assign each cell a cell cycle scores and visualize them in the DR plots.
We use known G2/M and S phase markers that come with the `Seurat` package. The markers are anticorrelated and cells that to not express the markers should be in G1 phase.

We compute cell cycle phase:
```{r cell-cycle}
# A list of cell cycle markers, from Tirosh et al, 2015
cc_file <- getURL("https://raw.githubusercontent.com/hbc/tinyatlas/master/cell_cycle/Homo_sapiens.csv") 
cc_genes <- read.csv(text = cc_file)
# match the marker genes to the features
m <- match(cc_genes$geneID[cc_genes$phase == "S"], 
           str_split(rownames(GetAssayData(so)),
                     pattern = "\\.", simplify = TRUE)[,1])
s_genes <- rownames(GetAssayData(so))[m]
(s_genes <- s_genes[!is.na(s_genes)])
m <- match(cc_genes$geneID[cc_genes$phase == "G2/M"], 
           str_split(rownames(GetAssayData(so)), 
                     pattern = "\\.", simplify = TRUE)[,1])
g2m_genes <- rownames(GetAssayData(so))[m]
(g2m_genes <- g2m_genes[!is.na(g2m_genes)])
so <- CellCycleScoring(so, s.features = s_genes, g2m.features = g2m_genes,
                       set.ident = TRUE)
```


# DR colored by cluster ID {.tabset}

```{r dim-red, results = "asis", fig.width = 12}
cs <- sample(colnames(so), 5e3)
.plot_dr <- function(so, dr, id)
    DimPlot(so, cells = cs, group.by = id, reduction = dr, pt.size = 0.4) +
        guides(col = guide_legend(nrow = 11, 
            override.aes = list(size = 3, alpha = 1))) +
        theme_void() + theme(aspect.ratio = 1)
ids <- c("cluster_id", "group_id", "sample_id", "Phase")
for (id in ids) {
    cat("## ", id, "\n")
    p1 <- .plot_dr(so, "tsne", id)
    p2 <- .plot_dr(so, "umap", id)
    if(id == "group_id") {
      p1 <- p1 + scale_color_manual(values = c("darkmagenta", "#7BAFDE", "#1965B0", "midnightblue"))
      p2 <- p2 + scale_color_manual(values = c("darkmagenta", "#7BAFDE", "#1965B0", "midnightblue"))
    }
    lgd <- get_legend(p1)
    p1 <- p1 + theme(legend.position = "none")
    p2 <- p2 + theme(legend.position = "none")
    ps <- plot_grid(plotlist = list(p1, p2), nrow = 1)
    p <- plot_grid(ps, lgd, nrow = 1, rel_widths = c(1, 0.2))
    print(p)
    cat("\n\n")
}
```

# Find markers using `scran`

We identify candidate marker genes for each cluster that enable a separation of that group from all other groups.

```{r scran-findMarkers, warning = FALSE}
scran_markers <- findMarkers(sce, 
    groups = sce$cluster_id, block = sce$sample_id, 
    direction = "up", lfc = 2, full.stats = TRUE)
```

## Heatmap of mean marker-exprs. by cluster
We aggregate the cells to pseudobulks and plot the average expression of the condidate marker genes in each of the clusters. 
```{r scran-markers-heatmap, fig.width = 12, fig.height = 6}
gs <- lapply(scran_markers, function(u) rownames(u)[u$Top == 1])
## candidate cluster markers
lapply(gs, function(x) str_split(x, pattern = "\\.", simplify = TRUE)[,2])
sub <- sce[unique(unlist(gs)), ]
pbs <- aggregateData(sub, assay = "logcounts", by = "cluster_id", fun = "mean")
mat <- t(muscat:::.scale(assay(pbs)))
## remove the Ensembl ID from the gene names
colnames(mat) <- str_split(colnames(mat), pattern = "\\.", simplify = TRUE)[,2] 
Heatmap(mat,
    name = "scaled avg.\nexpression",
    col = viridis(10),
    cluster_rows = FALSE,
    cluster_columns = FALSE,
    row_names_side = "left",
    row_title = "cluster_id",
    rect_gp = gpar(col = "white"))
```


# Known marker genes

```{r}
## source file with list of known marker genes
source(file.path("data", "known_cell_type_markers.R"))

fs <- lapply(fs, sapply, function(g)
    grep(pattern = paste0("\\.", g, "$"), rownames(sce), value = TRUE)
  )

fs <- lapply(fs, function(x) unlist(x[lengths(x) !=0]) )

gs <- gsub(".*\\.", "", unlist(fs))
ns <- vapply(fs, length, numeric(1))
ks <- rep.int(names(fs), ns)
labs <- lapply(fs, function(x) gsub(".*\\.", "",x))
```

## Heatmap of mean marker-exprs. by cluster

```{r known-markers-heatmap, fig.width = 10, fig.height = 18}
# split cells by cluster
cs_by_k <- split(colnames(sce), sce$cluster_id)
# compute cluster-marker means
ms_by_cluster <- lapply(fs, function(gs) vapply(cs_by_k, function(i)
        Matrix::rowMeans(logcounts(sce)[gs, i, drop = FALSE]), 
        numeric(length(gs))))
# prep. for plotting & scale b/w 0 and 1
mat <- do.call("rbind", ms_by_cluster)
mat <- muscat:::.scale(mat)
rownames(mat) <- gs
cols <- muscat:::.cluster_colors[seq_along(fs)]
cols <- setNames(cols, names(fs))
row_anno <- rowAnnotation(
    df = data.frame(label = factor(ks, levels = names(fs))),
    col = list(label = cols), gp = gpar(col = "white"))
# percentage of cells from each of the samples per cluster
sample_props <- prop.table(n_cells, margin = 1)
col_mat <- as.matrix(unclass(sample_props))
sample_cols <- c("#882255", "#CC6677", "#11588A", "#88CCEE", "#117733",  
                 "#44AA99",  "#FF8E21", "#FDC78E")

sample_cols <- setNames(sample_cols, colnames(col_mat))
col_anno <- HeatmapAnnotation(
    perc_sample = anno_barplot(col_mat, gp = gpar(fill = sample_cols), 
                               height = unit(2, "cm"),
                               border = FALSE),
    annotation_label = "fraction of sample\nin cluster",
    gap = unit(10, "points"))
col_lgd <- Legend(labels = names(sample_cols),
       title = "sample",
       legend_gp = gpar(fill = sample_cols))

hm <- Heatmap(mat,
    name = "scaled avg.\nexpression",
    col = viridis(10),
    cluster_rows = FALSE,
    cluster_columns = FALSE,
    row_names_side = "left",
    column_title = "cluster_id",
    column_title_side = "bottom",
    column_names_side = "bottom",
    column_names_rot = 0, 
    column_names_centered = TRUE,
    rect_gp = gpar(col = "white"),
    left_annotation = row_anno,
    top_annotation = col_anno)
draw(hm, annotation_legend_list = list(col_lgd))
```


# DR colored by marker expression {.tabset}

```{r dim-red-expr, results = "asis", fig.width = 12, fig.height = 12}
# downsample to 5000 cells
cs <- sample(colnames(sce), 5e3)
sub <- subset(so, cells = cs)
# UMAPs colored by marker-expression
for (m in seq_along(fs)) {
    cat("## ", names(fs)[m], "\n")
    ps <- lapply(seq_along(fs[[m]]), function(i) {
        if (!fs[[m]][i] %in% rownames(so)) return(NULL)
        FeaturePlot(sub, features = fs[[m]][i], reduction = "umap", pt.size = 0.4) +
            theme(aspect.ratio = 1, legend.position = "none") +
            ggtitle(labs[[m]][i]) + theme_void() + theme(aspect.ratio = 1)
    })
    # arrange plots in grid
    ps <- ps[!vapply(ps, is.null, logical(1))]
    p <- plot_grid(plotlist = ps, ncol = 4, label_size = 10)
    print(p)
    cat("\n\n")
}
```


# Marker genes for virus cell tropism

```{r read-virus-markers}
## source file with list of known marker genes
source(file.path("data", "virus_cell_tropism_markers.R"))
fs <- lapply(fs, sapply, function(g)
    grep(pattern = paste0("\\.", g, "$"), rownames(sce), value = TRUE)
  )

fs <- lapply(fs, function(x) unlist(x[lengths(x) !=0]) )

gs <- gsub(".*\\.", "", unlist(fs))
ns <- vapply(fs, length, numeric(1))
ks <- rep.int(names(fs), ns)
labs <- lapply(fs, function(x) gsub(".*\\.", "",x))
```

## Heatmap of virus cell tropism markers

```{r virus-markers-heatmap, fig.width = 9, fig.height = 10}
# compute cluster-marker means
ms_by_cluster <- lapply(fs, function(gs) vapply(cs_by_k, function(i)
        Matrix::rowMeans(logcounts(sce)[gs, i, drop = FALSE]), 
        numeric(length(gs))))
# prep. for plotting & scale b/w 0 and 1
mat <- do.call("rbind", ms_by_cluster)
mat <- muscat:::.scale(mat)
rownames(mat) <- gs
cols <- muscat:::.cluster_colors[seq_along(fs)]
cols <- setNames(cols, names(fs))
row_anno <- rowAnnotation(
    df = data.frame(label = factor(ks, levels = names(fs))),
    col = list(label = cols), gp = gpar(col = "white"))
# percentage of cells from each of the samples per cluster
sample_props <- prop.table(n_cells, margin = 1)
col_mat <- as.matrix(unclass(sample_props))
sample_cols <- c("#882255", "#CC6677", "#11588A", "#88CCEE", "#117733",  
                 "#44AA99",  "#FF8E21", "#FDC78E")
sample_cols <- setNames(sample_cols, colnames(col_mat))
col_anno <- HeatmapAnnotation(
    perc_sample = anno_barplot(col_mat, gp = gpar(fill = sample_cols), 
                               height = unit(2, "cm"),
                               border = FALSE),
    annotation_label = "fraction of sample\nin cluster",
    gap = unit(10, "points"))
col_lgd <- Legend(labels = names(sample_cols),
       title = "sample",
       legend_gp = gpar(fill = sample_cols))

hm <- Heatmap(mat,
    name = "scaled avg.\nexpression",
    col = viridis(10),
    cluster_rows = FALSE,
    cluster_columns = FALSE,
    row_names_side = "left",
    column_title = "cluster_id",
    column_title_side = "bottom",
    column_names_side = "bottom",
    column_names_rot = 0, 
    column_names_centered = TRUE,
    rect_gp = gpar(col = "white"),
    left_annotation = row_anno,
    top_annotation = col_anno)
draw(hm, annotation_legend_list = list(col_lgd))
```

# Plot for paper

We want to find more markers specific for the NSCs:
```{r NSC-markers}
## markers specific for cluster 0, 1 or 14 (NSC) versus all other clusters
scran_markers2 <- findMarkers(sce, groups = sce$cluster_id, 
                              direction = "up", lfc = 2, pval.type="all")
scran_markers2[["0"]][1:10,1:3]
scran_markers2[["1"]][1:10,1:3]
scran_markers2[["14"]][1:10,1:3]
## we selected following markers
g <- c("NQO1", "TPM1", "PEG10")
genes <- lapply(g, sapply, function(i)
    grep(pattern = paste0("\\.", i, "$"), rownames(sce), value = TRUE)
  )
names(genes) <- g

# UMAPs colored by marker-expression
for (m in seq_along(genes)) {
  p <- FeaturePlot(sub, features = genes[[m]], reduction = "umap", pt.size = 0.4) +
            theme(aspect.ratio = 1, legend.position = "none") +
            ggtitle(names(genes)[[m]]) + theme_void() + theme(aspect.ratio = 1)
    print(p)
}
```

We plot the gene expression of some selected markers for the paper.
```{r paper-figures, fig.width = 9, fig.height = 9}
#  SOX2 and NQO1 (NSCs), GFAP (astrocytic), PTPRZ1 (OPCs), DCN (pericytes), 
#  SYP (neuronal), SLC32A1 (GABAergic) or SLC17A6 (Glutamatergic)
g <- c("SOX2","NQO1",  "GFAP", "PTPRZ1", "DCN", "SYP", "SLC32A1", "SLC17A6")
genes <- lapply(g, sapply, function(i)
    grep(pattern = paste0("\\.", i, "$"), rownames(sce), value = TRUE)
  )
names(genes) <- g

ps <- lapply(names(genes)[1:2], function(i) {
        FeaturePlot(sub, features = genes[[i]], reduction = "umap", pt.size = 0.4,
                    max.cutoff = 3) +
            ggtitle(i) + theme_void() + theme(aspect.ratio = 1,
                                              plot.title = element_text(hjust = 0.5))
    })
  lgd1 <- get_legend(ps[[1]])
  ps <- lapply(ps, function(x) x + theme(legend.position = "none"))
  ps1 <- plot_grid(plotlist = ps, nrow = 1)
  p1 <- plot_grid(ps1, lgd1, nrow = 1, label_size = 10, rel_widths = c(1, 0.1))
    
 ps <- lapply(names(genes)[3:5], function(i) {
        FeaturePlot(sub, features = genes[[i]], reduction = "umap", pt.size = 0.4,
                    min.cutoff = 0, max.cutoff = 3.5) +
            ggtitle(i) + theme_void() + theme(aspect.ratio = 1,
                                              plot.title = element_text(hjust = 0.5))
    })
  lgd2 <- get_legend(ps[[1]])
  ps <- lapply(ps, function(x) x + theme(legend.position = "none"))
  ps2 <- plot_grid(plotlist = ps, nrow = 1)
  p2 <- plot_grid(ps2, lgd2, nrow = 1, label_size = 10, rel_widths = c(1, 0.1))
  p <- plot_grid(p1, p2, nrow = 2)
  
 ps <- lapply(names(genes)[6:8], function(i) {
        FeaturePlot(sub, features = genes[[i]], reduction = "umap", pt.size = 0.4,
                    min.cutoff = 0, max.cutoff = 2.5) +
            ggtitle(i) + theme_void() + theme(aspect.ratio = 1,
                                              plot.title = element_text(hjust = 0.5))
    })
  lgd3 <- get_legend(ps[[1]])
  ps <- lapply(ps, function(x) x + theme(legend.position = "none"))
  ps3 <- plot_grid(plotlist = ps, nrow = 1)
  p3 <- plot_grid(ps3, lgd3, nrow = 1, label_size = 10, rel_widths = c(1, 0.1))
  p <- plot_grid(p1, p2, p3, nrow = 3)
  
  print(p)
```


How are the UNC13A expression levels in the different clusters and time points?

```{r UNC13A-expression}
## we selected following markers
g <- "UNC13A"
gene <- grep(pattern = paste0("\\.", g, "$"), rownames(sce), value = TRUE)
names(gene) <- g

# UMAP colored by marker-expression
p <- FeaturePlot(sub, features = gene, reduction = "umap", pt.size = 0.4) +
          theme(aspect.ratio = 1, legend.position = "none") +
          ggtitle(g) + theme_void() + theme(aspect.ratio = 1)
print(p)
```


```{r UNC13A-heatmap, fig.width = 9, fig.height = 3}
# compute cluster-marker means
mat <- vapply(cs_by_k, function(i)
        Matrix::rowMeans(logcounts(sce)[gene, i, drop = FALSE]), 
        numeric(length(gene)))
# prep. for plotting & scale b/w 0 and 1
scale_vector <- function (x) {
    qs <- rowQuantiles(t(as.matrix(x)), probs = c(0.01, 0.99), na.rm = TRUE)
    x <- (x - qs[1])/(qs[2] - qs[1])
    x[x < 0] <- 0
    x[x > 1] <- 1
    return(t(as.matrix(x)))
}

mat <- scale_vector(mat)
# percentage of cells from each of the samples per cluster
sample_props <- prop.table(n_cells, margin = 1)
col_mat <- as.matrix(unclass(sample_props))
sample_cols <- c("#882255", "#CC6677", "#11588A", "#88CCEE", "#117733",  
                 "#44AA99",  "#FF8E21", "#FDC78E")
sample_cols <- setNames(sample_cols, colnames(col_mat))
col_anno <- HeatmapAnnotation(
    perc_sample = anno_barplot(col_mat, gp = gpar(fill = sample_cols), 
                               height = unit(2, "cm"),
                               border = FALSE),
    annotation_label = "fraction of sample\nin cluster",
    gap = unit(10, "points"))
col_lgd <- Legend(labels = names(sample_cols),
       title = "sample",
       legend_gp = gpar(fill = sample_cols))

hm <- Heatmap(mat,
    name = "scaled avg.\nexpression",
    col = viridis(10),
    cluster_rows = FALSE,
    cluster_columns = FALSE,
    row_names_side = "left",
    column_title = "cluster_id",
    column_title_side = "bottom",
    column_names_side = "bottom",
    column_names_rot = 0, 
    column_names_centered = TRUE,
    rect_gp = gpar(col = "white"),
    # left_annotation = row_anno,
    top_annotation = col_anno)
draw(hm, annotation_legend_list = list(col_lgd))
```
